Convert to using GNU autotest tool
The testsuite support in automake is very limited and inadequate for purpose.
The GNU autotest tool is better able to satisfy the requiremtns of running
different tests differently even in a cross-build environment. It also supports
the ability to install or package the testsuite separately so it can be run on a
remote target.

Because of the increased flexibility of the GNU autotest tool, a number of test
runner scripts were no longer necessary and have been removed.

The github action was updated ot use thw new test tool instead of the
automake-generated one.